first pass at reorganising aotools
