+++ aotools/__init__.py
@@ -1,8 +1,6 @@
 from .aotools import *
 
-from . import wfs, circle, fft, interp, turbulence
-
-phasescreen = turbulence # For compatibility
+from . import astronomy, fft, functions, image_processing, interp, turbulence, wfs
 
 from ._version import get_versions
 __version__ = get_versions()['version']
