+++ aotools/astronomy/astronomy.py
@@ -1,7 +1,3 @@
-"""
-Unspecified tools for AO and astronomy
-"""
-
 def photonsPerMag(mag, mask, pxlScale, wvlBand, expTime):
     '''
     Calculates the photon flux for a given aperture, star magnitude and wavelength band
@@ -30,36 +26,3 @@ def photonsPerMag(mag, mask, pxlScale, wvlBand, expTime):
     photons = photonPerSec * expTime
 
     return photons
-
-def image_contrast(image):
-    """
-    Calculates the 'Michelson' contrast.
-
-    Uses a method by Michelson (Michelson, A. (1927). Studies in Optics. U. of Chicago Press.), to calculate the contrast ratio of an image. Uses the formula:
-        (img_max - img_min)/(img_max + img_min)
-
-    Parameters:
-        image (ndarray): Image array
- 
-    Returns:
-        float: Contrast value
-    """
-
-    contrast = (image.max() - image.min()) / (image.max() + image.min())
-
-    return contrast
-
-def rms_contrast(image):
-    """
-    Calculates the RMS contrast - basically the standard deviation of the image
-
-    Parameters:
-        image (ndarray): Image array
-
-    Returns:
-        float: Contrast value
-    """
-    
-    image /= image.max()
-    
-    return image.std()
