--- aotools/circle/__init__.py
@@ -1,2 +1,0 @@
-from .circle import *
-from .zernike import *
