+++ aotools/turbulence/__init__.py
@@ -1,1 +1,4 @@
 from .phasescreen import *
+from .infinitephasescreen import *
+from .temporal_ps import *
+from .opticalpropagation import *
